requirements.txt
scikit-learn
Pillow
numpy
rich
